fix: correct type definitions for startView and stopView

+++ src/types/native-playout-audio-worklet-node-options.ts
@@ -6,9 +6,9 @@ export type TNativePlayoutAudioWorkletNodeOptions = Partial<Omit<TNativeAudioWor
 
     readPointerView: Uint8Array | Uint16Array | Uint32Array;
 
-    startView: Uint8Array;
+    startView: Uint16Array;
 
-    stopView: Uint8Array;
+    stopView: Uint16Array;
 
     storageView: Float32Array;
